在spi.c中增加CSN_1和CSN_2的配置，可以分开控制

+++ sdk/peripheral/adc.c
@@ -39,7 +39,7 @@ AdcTypedef adcstruct;
 
 
 
-void adc_delay_us(float dly1us)
+void adc_delay_us(uint32 dly1us)
 {
 	volatile uint32 jj=0;
 	while(dly1us--)
